Bug 1233798 - report to console when service worker register fails due to mime-type issues. r=bkelly
Log a message to the console when there is no Content-Type received for a
service worker or it's not a JS type.  With test.  Looks like:

Failed to register/update a ServiceWorker: Non-JS Content-Type of `text/plain'
received for script
`http://mochi.test:8888/tests/dom/workers/test/serviceworkers/sw_bad_mime_type.js'.