Bug 1256482 - Disable C4312 to unblock compilation on VS2015; r=jesup
As part of unblocking building with VS2015u1 in automation, I'm mass
disabling compiler warnings that are turned into errors. This is not
the preferred mechanism to fix compilation warnings, of course.

Given that the warning is in WebRTC which is third party code, there
isn't much we can do about the warning. However, Google is building
Chrome with Visual Studio 2015, so I wouldn't be surprised if this
were fixed upstream or will be fixed upstream. Then again, we allow
warnings when building WebRTC. So perhaps not.