Bug 1264566 - Part 1: Refactor FileDescriptor to fix leaks in content processes. r=valentin, r=baku
Now FileDescriptor takes the ownership of the platform handle, in both parent
and content processes.
Copying will duplicate the underlying handle.
It also comes with a move constructor to avoid duplicating if possible.

The getter will duplicate a new handle, so a UniquePtr is needed to hold the
duplicated handle.