Bug 1254070 - inspector: trigger autocomplete if nextchar is not a word character;r=gl
For the style attribute CSS autocompletion, suggestions will now be provided
when the character after the caret is a non-word character. As opposed to the
current behavior where suggestions would only be displayed if the next character
was the end of the input or a space.

This allows to show suggestions when the user's caret is next to ";" or '"', which
can frequently happen when editing style attributes.