Bug 1279135 - Reattach Pipeline to PipelineTransport in AttachTransport_s; r=jesup
When we call MediaPipeline::UpdateTransport_s we in turn call DetachTransport_s
which detaches the pipeline from PipelineTransport. The subsequent call to
AttachTransport_s does not currently reattach the pipeline, causing
subsequent sends to fail due to a detached pipeline. Since
PipelineTransport::SendRtpRtcpPacket_s returns NS_OK if a send fails due to a
detached pipeline, this failure is not straightforward to detect.

This patch adds an Attach() method to PipelineTransport and calls it from
AttachTransport_s.