Bug 1128069: [MSE] P6. Call NotifyDataArrived from MediaDecoder. r=gerald
Calling NotifyDataArrived from each sourcebuffer will cause multiple unnecessary NotifyDataArrived to the MediaFormatReader when it could only be done once. Additionally, it ensures that the media duration is updated prior to the reader actioning on the notification.

Extra: mEnded is only ever accessed on the main thread, there's no need to make it atomic.