Bug 1278612 - Use DefaultOutputUnit as the default output device.  r=jesup
kinetiknz/cubeb@6b2c610 changed the output unit
from kAudioUnitSubType_DefaultOutput to kAudioUnitSubType_HALOutput because
capture is never available on the DefaultOutputUnit.  For the case where
we're doing output only to the default device, this regressed the automatic
device switching when the output device was changed in the Sound system
preferences.  Reverting to the DefaultOutputUnit for this case restores the
previous behaviour.  This addresses BMO #1278612.