Bug 1288870, part 2 - Handlify the argument to XPCNativeSet::FindMember(). r=mrbkap
This is to work around false GC hazards in ~XPCNativeInterface in part
4 of this bug. Putting RefPtr<XPCNativeInterface> on the stack means
that ~XPCNativeInterface can get called in various places, and the GC
hazard analysis does not understand the virtual methods involved, so
it assumes they might possibly GC.

This fixes one hazard by taking a jsid handle argument. The callers
already pass in handles, so no other changes are needed.